requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.13
PyMuPDF==1.23.8
python-docx==1.1.0
chromadb==0.4.22
sentence-transformers==5.1.0
python-dotenv==1.0.0
tiktoken==0.5.2
groq==0.32.0